Address Clippy lints

+++ graphics/src/widget/qr_code.rs
@@ -67,10 +67,7 @@ where
         let side_length = (self.state.width + 2 * QUIET_ZONE) as f32
             * f32::from(self.cell_size);
 
-        layout::Node::new(Size::new(
-            f32::from(side_length),
-            f32::from(side_length),
-        ))
+        layout::Node::new(Size::new(side_length, side_length))
     }
 
     fn draw(
